+++ ownership-chain/node/src/command.rs
@@ -42,9 +42,9 @@ fn load_spec(id: &str) -> std::result::Result<Box<dyn ChainSpec>, String> {
 		"giedi" => Box::new(chain_spec::ChainSpec::from_json_bytes(
 			&include_bytes!("../../specs/giedi.raw.json")[..],
 		)?),
-		"dev" => Box::new(chain_spec::development_config()),
-		"template-rococo" => Box::new(chain_spec::local_testnet_config()),
-		"" | "local" | "local-v" => Box::new(chain_spec::local_testnet_config()),
+		"dev" => Box::new(chain_spec::laos::development_config()),
+		"template-rococo" => Box::new(chain_spec::laos::local_testnet_config()),
+		"" | "local" | "local-v" => Box::new(chain_spec::laos::local_testnet_config()),
 		path => Box::new(chain_spec::ChainSpec::from_json_file(std::path::PathBuf::from(path))?),
 	})
 }
